implementing loadFlashcards(), updateCardData(), and BookActivity.java's two FABs

+++ app/src/main/java/edu/ucdenver/sweanyn/flashcardfinal/FlashcardDao.java
@@ -16,7 +16,7 @@ public interface FlashcardDao {
     List<Flashcard> getAll();
 
     @Query("SELECT * FROM flashcards WHERE book_id = :bookId")
-    List<Flashcard> getByBookId(int bookId);
+    List<Flashcard> getByBookId(long bookId);
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
     long insert(Flashcard flashcard);
